v1

+++ backend/track-service/internal/domain/model/model.go
@@ -2,7 +2,7 @@ package model
 
 // Track represents a track in the system
 type Track struct {
-	ID       int    `json:"id" db:"id" redis:"id"`
+	ID       int64  `json:"id" db:"id" redis:"id"`
 	Title    string `json:"title" db:"title" redis:"title"`
 	Duration int    `json:"duration" db:"duration" redis:"duration"`
 	AudioURL string `json:"audio_url" db:"audio_url" redis:"audio_url"`
